move add exception hook, lint
