+++ x-pack/plugins/security_solution/public/common/components/exceptions/add_exception_modal/index.tsx
@@ -33,7 +33,7 @@ import { useKibana } from '../../../lib/kibana';
 import { errorToToaster, displaySuccessToast, useStateToaster } from '../../toasters';
 import { ExceptionBuilder } from '../builder';
 import { Loader } from '../../loader';
-import { useAddOrUpdateException } from '../../../../alerts/containers/detection_engine/alerts/use_add_exception';
+import { useAddOrUpdateException } from '../use_add_exception';
 import { useSignalIndex } from '../../../../alerts/containers/detection_engine/alerts/use_signal_index';
 import { useFetchOrCreateRuleExceptionList } from '../use_fetch_or_create_rule_exception_list';
 import { AddExceptionComments } from '../add_exception_comments';
